Add a tar of the code and a rsync .deb file
The tar is for loading directly on a Pi via a USB stick. The files are
in Unix format, so they will work in the Pi environment.

The rsync .deb file needs to be installed before running rpi-clone if
the Pi is not connected to the internet.  Install with:

dpkg -i rsync_3.1.3-6_armhf.deb