Support Ubuntu by modifying /boot/firmware/cmdline.txt
On Ubuntu, cmdline.txt is stored in a subdirectory, so if the file does
not exist in the usual location, fall back to the Ubuntu location (and
fall back to the Armbian version if the Ubuntu version does not exist
either).

Note that this does not support Ubuntu completely out of the box, since
it uses filesystem labels instead of UUIDs for mounting, but if you
manualy change them to UUIDs, this change allows cloning to work.