Support Armbian by modifying armbianEnv.txt
On Armbian, the kernel commandline (or at least the root device to use)
is stored in armbianEnv.txt rather than cmdline.txt, so update that if
it exists and cmdline.txt does not.

The actual operation on this file is just replacing UUIDs inside the
file, so it works even though the syntax of the file is different.