Refactor cmdline.txt with more variables
This removes all hardcoded references to cmdline.txt and cmdline.boot by
removing the `$clone` prefix from cmdline_txt and introducing a
cmdline_txt variable.

In addition, this adds some quoting to any lines modified to be a bit
more robust when spaces are involved.

This should not change behavior (apart from maybe in cases with spaces
that were previously broken), the commands ran should be identical.