Refactor cmdline/fstab updating into functions
This mostly moves code into functions, but also makes minor changes to
how the code is called to prepare for more reuse of this code.

Moving these functions up in the file with the other functions is left
for the next commit to make review easier.

Diff best viewed with --ignore-all-space.