Simplify fixup_cmdline_txt()
This applied largely the same operations to cmdline.txt as
fixup_device_references_in_file did to fstab, so this can just call the
latter.

This should work the same, except that the PARTUUID is not replaced with
the g flag (all occurences instead of just one), all filesystem UUIDs
are replaced (instead of just the root partition), and progress output
messages are slightly different.

Also, fstab UUID= replacements are now done anywhere in the file,
instead of just at the start of a line.