Improve output for cmdline/fstab updates
This modifies fixup_device_references_in_file to accept the mountpoint
and filename separately to prevent printing the destination mountpoint
(which might be ugly and meaningless). fixup_boot_partition already had
this.

Then, a partition label is passed which is prefixed to this plain
filename to tell the user which partition is being modified. This could
just have used src_mount, but passing a separate label prepares for
using this code for non-mounted partitions too.

Finally, the UUID output is slightly reworded and some indentation is
added (preparing for this output to also be intermixed with mounting and
rsync calls later).